+++ src/main/java/tech/anonymoushacker1279/iwcompatbridge/plugin/lucent/LucentPluginHandler.java
@@ -3,9 +3,7 @@ package tech.anonymoushacker1279.iwcompatbridge.plugin.lucent;
 import com.legacy.lucent.api.EntityBrightness;
 import com.legacy.lucent.api.plugin.ILucentPlugin;
 import com.legacy.lucent.api.plugin.LucentPlugin;
-import tech.anonymoushacker1279.immersiveweapons.entity.projectile.MortarShellEntity;
-import tech.anonymoushacker1279.immersiveweapons.entity.projectile.SmokeGrenadeEntity;
-import tech.anonymoushacker1279.immersiveweapons.entity.projectile.bullet.BulletEntity;
+import tech.anonymoushacker1279.immersiveweapons.entity.projectile.*;
 import tech.anonymoushacker1279.iwcompatbridge.IWCompatBridge;
 import tech.anonymoushacker1279.iwcompatbridge.config.CommonConfig;