+++ src/main/java/tech/anonymoushacker1279/iwcompatbridge/plugin/wthit/components/TeslaSynthesizerComponent.java
@@ -12,12 +12,14 @@ public class TeslaSynthesizerComponent implements IBlockComponentProvider {
 			int cookTime = accessor.getData().raw().getInt("CookTime");
 			int cookTimeTotal = accessor.getData().raw().getInt("CookTimeTotal");
 
-			// Calculate remaining time
-			int remainingTime = cookTimeTotal - cookTime;
-			int minutes = remainingTime / 1200;
-			int seconds = (remainingTime % 1200) / 20;
+			if (cookTime > 0) {
+				// Calculate remaining time
+				int remainingTime = cookTimeTotal - cookTime;
+				int minutes = remainingTime / 1200;
+				int seconds = (remainingTime % 1200) / 20;
 
-			tooltip.addLine(Component.translatable("iwcompatbridge.wthit.tesla_synthesizer.progress", minutes, seconds));
+				tooltip.addLine(Component.translatable("iwcompatbridge.wthit.tesla_synthesizer.progress", minutes, seconds));
+			}
 		}
 	}
 }