+++ RELEASE_NOTES.md
@@ -1,9 +1,9 @@
-This release supports IW v1.26.0.
+This release supports IW v1.27.0.
 
 ### New Features
 
-- Cook time for the Tesla Synthesizer now displayed via WTHIT
+- JEI support for the Ammunition Table
 
 ### Bugfixes
 
-- Update category in PMMO plugin to be `crafting` instead of `craft`
+- Tesla Synthesizer cook time is no longer displayed at 0 seconds